+++ cs-20151215/ChangeLog.md
@@ -1,3 +1,8 @@
+2026-05-15 Version: 6.6.3
+- Update API DescribeClusterNodePoolDetail: add response parameters Body.eflo_node_group.
+- Update API DescribeClusterNodePools: add response parameters Body.nodepools.$.eflo_node_group.
+
+
 2026-05-12 Version: 6.6.1
 - Update API DescribeClusterEvents: add request parameters max_results.
 - Update API DescribeClusterEvents: add request parameters next_token.